[fix] write, update post

+++ DDW/src/main/resources/static/js/app/board.js
@@ -1,6 +1,6 @@
 var main = {
     init : function () {
-        var _this = this;
+        const _this = this;
         $('#btn-save').on('click', function () {
             _this.save();
         });
@@ -14,7 +14,7 @@ var main = {
         });
     },
     save : function () {
-        var data = {
+        const data = {
             title: $('#title').val(),
             owner: $('#owner').val(),
             content: $('#content').val()
@@ -34,40 +34,53 @@ var main = {
         });
     },
     update : function () {
-        var data = {
+        const data = {
             title: $('#title').val(),
             content: $('#content').val()
         };
 
-        var id = $('#id').val();
+        const id = $('#id').val();
+        const con_check = confirm("수정하시겠습니까?");
+         if (con_check === true) {
+            if (!data.title || data.title.trim() === "" || !data.content || data.content.trim() === "") {
+                alert("제목과 내용을 모두 입력해 주세요.");
+                return false;
+            } else {
+                    $.ajax({
+                        type: 'PUT',
+                        url: '/api/v1/posts/'+id,
+                        dataType: 'json',
+                        contentType:'application/json; charset=utf-8',
+                        data: JSON.stringify(data)
+                    }).done(function() {
+                        alert('글이 수정되었습니다.');
+                        window.location.href = '/board/list';
+                    }).fail(function (error) {
+                        alert(JSON.stringify(error));
+                 });
+             }
+         }
+     },
+     delete : function () {
+        const id = $('#id').val();
+        const con_check = confirm("정말 삭제하시겠습니까?");
 
-        $.ajax({
-            type: 'PUT',
-            url: '/api/v1/posts/'+id,
-            dataType: 'json',
-            contentType:'application/json; charset=utf-8',
-            data: JSON.stringify(data)
-        }).done(function() {
-            alert('글이 수정되었습니다.');
-            window.location.href = '/board/list';
-        }).fail(function (error) {
-            alert(JSON.stringify(error));
-        });
-    },
-    delete : function () {
-        var id = $('#id').val();
+        if(con_check == true) {
+            $.ajax({
+                type: 'DELETE',
+                url: '/api/v1/posts/'+id,
+                dataType: 'JSON',
+                contentType: 'application/json; charset=utf-8'
 
-        $.ajax({
-            type: 'DELETE',
-            url: '/api/v1/posts/'+id,
-            dataType: 'json',
-            contentType:'application/json; charset=utf-8'
-        }).done(function() {
-            alert('글이 삭제되었습니다.');
-            window.location.href = '/board/list';
-        }).fail(function (error) {
-            alert(JSON.stringify(error));
-        });
+            }).done(function () {
+                alert("삭제되었습니다.");
+                window.location.href = '/board/list';
+            }).fail(function (error) {
+                alert(JSON.stringify(error));
+            });
+        } else {
+            return false;
+        }
     }
 
 };
